Django Backend/MyUsers/requirements.txt
Django==4.0.3
django_rest_knox==4.2.0
djangorestframework==3.13.1
knox==0.1.14
requests==2.21.0
Pillow==10.1.0
